requirements.txt
pysdl2
pysdl2-dll
pyside6
pyinstaller